+++ src/bot/commands/music/spotify/artist.ts
@@ -76,42 +76,40 @@ export default class SpotifyArtistCommand extends Command {
         this.client.spotify.getArtist(artistId).then(async (res) => {
           const artistEmbed = new MessageEmbed();
           const userAgent = 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) '
-            + 'Chrome/78.0.3879.0 Safari/537.36 Edg/78.0.249.0';
+            + 'Chrome/78.0.3886.0 Safari/537.36 Edg/78.0.257.0';
           const accessTokenUrl = await request.get('https://open.spotify.com').set({ 'User-Agent': userAgent });
           // TODO: Store token and only refresh the token when it expires.
           const accessToken = accessTokenUrl.header['set-cookie'][3].split('=')[1].split(';')[0];
 
-          const artistAbout = await request.get(spotifyBackendFullUrl + artistId).set({
+          const about = await request.get(spotifyBackendFullUrl + artistId).set({
             Authorization: 'Bearer ' + accessToken,
             'User-Agent': userAgent,
           });
 
-          // Artist Information
-          let artistBiography: string;
-          const artistName = res.body.name;
-          const artistGenres = res.body.genres.join(', ');
-          const artistLink = res.body.external_urls.spotify;
-          const artistImage = res.body.images[0].url;
-          // Statistics
-          const artistFollowers = res.body.followers.total;
-          const artistListeners = artistAbout.body.artistInsights.monthly_listeners;
+          let biography: string;
+          const name = res.body.name;
+          const genres = res.body.genres.join(', ');
+          const link = res.body.external_urls.spotify;
+          const image = res.body.images[0].url;
+          const followers = res.body.followers.total;
+          const listeners = about.body.artistInsights.monthly_listeners;
 
-          if (artistAbout.body.hasOwnProperty('bio')) {
-            artistBiography = stripHtml(Util.shorten(artistAbout.body.bio, 1000));
+          if (about.body.hasOwnProperty('bio')) {
+            biography = stripHtml(Util.shorten(about.body.bio, 1000));
           } else {
-            artistBiography = 'No biography available.';
+            biography = 'No biography available.';
           }
 
-          artistEmbed.setTitle(artistName);
+          artistEmbed.setTitle(name);
           artistEmbed.setColor(0x1DB954);
-          artistEmbed.setURL(artistLink);
-          artistEmbed.setThumbnail(artistImage);
+          artistEmbed.setURL(link);
+          artistEmbed.setThumbnail(image);
           artistEmbed.setDescription(
-            `${artistBiography}\n\n` +
+            `${biography}\n\n` +
             '**__Artist Stats__:**\n' +
-            `**Followers**: ${artistFollowers}\n` +
-            `**Listeners**: ${artistListeners ? artistListeners : `No users listen to ${artistName}.`}\n` +
-            `**Genres**: ${artistGenres ? artistGenres : 'No genres available.'}\n\n`);
+            `**Followers**: ${followers}\n` +
+            `**Listeners**: ${listeners ? listeners : `No users listen to ${name}.`}\n` +
+            `**Genres**: ${genres ? genres : 'No genres available.'}\n\n`);
 
           return message.channel.send(artistEmbed);
         });